Add --time parameter to override default time calculation
Needed to ensure that tfb inside the docker container uses
the folder that we expect for results.

If we don't force the timestamp, then you end up with a results folder for
the tfb running inside docker and a results folder for the tfb running
outside docker, as their respective runtimes are different.

The folder created by tfb inside the docker container will have raw
results, while the one created outside the docker container will have only
the results.json file