Ensure files/folders created inside docker container can be accessed outside
Because TFB runs as root inside the docker container, any folders or files
that are created inside the container can not be written to by the TFB
process running outside the docker container. Practically this means that
the parent TFB process cannot read write results.json inside the
<test>/<time> folder

The solution for now is to chmod 777 files when they are created inside the
container. The long term solution would be to create a user inside the container
that corresponds to a user outside the container and then run TFB as that
user inside the container.